fix: reorganize folder structure
