+++ static/assets.txt
