+++ static/config.js
