+++ templates/coloradoskyrunner.html
@@ -5,7 +5,7 @@
     <title>Colorado Sky Runner</title>
     <style>
       body {
-        margin: R0;
+        margin: 0;
         padding: 0;
         background-color: #000;
         overflow: hidden;
@@ -85,7 +85,7 @@
   </head>
   <body>
     <div id="container">
-      <img id="logo" src="./logo.png" alt="Logo">
+      <img id="logo" src="/static/logo.png" alt="Logo">
       <div id="score">Score: 0</div>
       <div id="ui">
         <button id="aboutBtn">About</button>
@@ -106,6 +106,6 @@
         }
       }
     </script>
-    <script type="module" src="main.js"></script>
+    <script type="module" src="/static/main.js"></script>
   </body>
 </html>
